fix: fix middleware auth checks
Middleware needed to be moved to the src folder to be recognized.
Authentication for /home was failing because of that and the tests were failing.
Fix the issue and changed the tests to check that we are being redirected to the login page.
Also updated the README.md to be more specific about what's required.